requirements.txt
ultralytics==8.1.0
clearml
torch==2.1.0+cu121
torchvision==0.16.0+cu121
torchaudio==2.1.0+cu121
numpy
pyyaml
